+++ src/Frontend/Dfe.Complete.Logging/Models/ErrorResponse.cs
@@ -1,6 +1,6 @@
-using System.Text.Json;
+﻿using System.Text.Json; 
 
-namespace Dfe.Complete.Api.ResponseModels
+namespace Dfe.Complete.Logging.Models
 {
     public class ErrorResponse
     {